refactor(EIDOMNI-844): Split Trust Protocol 1.0 logic into its own service class
