+++ verifier-service/src/test/java/ch/admin/bj/swiyu/verifier/service/oid4vp/DcqlVpTokenVerifierTest.java
@@ -4,6 +4,7 @@ import ch.admin.bj.swiyu.verifier.common.exception.VerificationException;
 import ch.admin.bj.swiyu.verifier.domain.SdJwt;
 import ch.admin.bj.swiyu.verifier.domain.management.Management;
 import ch.admin.bj.swiyu.verifier.domain.management.dcql.DcqlCredential;
+
 import com.authlete.sd.Disclosure;
 import com.nimbusds.jwt.JWTClaimsSet;
 import org.junit.jupiter.api.BeforeEach;
