+++ verifier-service/src/test/java/ch/admin/bj/swiyu/verifier/service/oid4vp/service/IssuerTrustValidatorTest.java
@@ -6,6 +6,7 @@ import ch.admin.bj.swiyu.verifier.domain.management.TrustAnchor;
 import ch.admin.bj.swiyu.verifier.service.oid4vp.IssuerTrustValidator;
 import ch.admin.bj.swiyu.verifier.service.oid4vp.SdJwtVpTokenVerifier;
 import ch.admin.bj.swiyu.verifier.service.publickey.IssuerPublicKeyLoader;
+
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
